use single connection to a random database server per request
Previously each database server on Tool Labs (and Toolserver before that) only had a subset of the databases, so crosswiki tools needed to maintain multiple connections to specific DB servers. With recent changes to Tool Labs, each server has a copy of every database. This commit makes tools open a single connection per request to a random database server. This reduces the dependence on any one server -- if a page breaks because of a DB server issue, the next refresh will probably get a different server.